Restore and remember styling during selection
